Initial implementation complete, is mostly working
All parameter scaling and such seems to be correct, and the hi-pass
filter definitely works. The lo-pass filter seems to be broken for some
strange reason, though.

Anyways, the filter action is a bit crackly when testing against the
A440 wave, but is otherwise acceptable. Performance is very nice as
well, with 60+ instances using ~17% CPU on my MacBook Pro. The same
number of instances of "The Pilgrim" (which inspired this plugin) uses
nearly twice as much CPU.